Switch the capabilities to a read/write mode
Instead of separate capabilities flags for every combination, support a 'r' or
'w' mode for each capability. This way we can easily communicate that a device
supports writing configuration but not reading it (steelseries) or that we can
e.g. read a LED but not write to it.

This means we can drop some of the current enums, writable
profiles/resolutions now implies switchable profiles/resolutions until we find
a device where this isn't true and we need an extra flag.

The rest is mostly search-replace. Note that the ratbagd API is still as-is
here with a single Capabilities property exposed - currently set to the
read capabilities.

Fixes #265

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>