ratbagd: expose the capabilities as read/write with the corresponding mode
The new content of the Capabilities property is now a list of tuples with the
capability as first entry and the modes as second entry as string, e.g.

 (1, "rw"), (2, "r"), (3, "w"), (4, "wr")

In the future extra modes may need to get added, the string could thus become
"rwa" or something like that too.

This is a DBus API break.

Fixes #265

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>