hidpp20: Extend sensor information to handle X/Y/LOD
The existing hidpp20 code base handles only a singular DPI value to
support the 2201 Adjustable DPI feature alongside onboard profiles which
match. Newer devices (e.g. Logitech G Pro X Superlight 2) add support
for per axis dpi configuration as well as Lift off distance.

This change reworks the existing hidpp20/driver-hidpp20 to describe
sensors that are capable of dual axis DPI and lift off distance. Changes
to the hidpp20_sensor structure provide per axis information in new
hidpp20_sensor_axis structures along with fixed details for the lift off
distance. For devices that only have one axis of configuration, the X
axis is used to represent the singular value/capability (which matches
other drivers in ratbag).

Additionally the newer devices are able to provide multiple DPI ranges
with different step increments per range e.g. 100-200 in steps 1
followed by 200-400 in steps of 2. The hidpp20_sensor_axis structure
includes the DPI ranges with an array of a hidpp20_sensor_dpi_range
structures. This allows for capturing the >1000 distinct dpi values
without storing >1000 values. The hidpp20drv_validate_dpi_ranges
function is added in order to handle the validation of DPI values used
in various setters. The DPI list values provided to the upper ratbag API
is still generated by standard values in the dpi_min/dpi_max range, this
is because the ratbag API does not allow for the >1000 values to be
stored (and the resolution of those rates are not particularly useful).

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>