hidpp20: Add support for onboard profile format id 0x06
Newer devices including the Logitech G Pro X Wireless Superlight 2
implement onboard profiles with a new format id of 0x06. This format is
significantly different from the existing formats due to the support for
features 0x2202 and 0x8061 which includes support for independent axis
DPI values, Lift Off Distance (LOD) and extended report rates including
wired/wireless specific values.

This change adds support for the basic functionality that is present in
this profile format, as the G Pro X Wireless Superlight 2 does not
provide some of the extra functions that are present in the format
profile structure (e.g. LEDs). This profile format also provide
configuration specifics that are not currently possible with the ratbag
API.

The X and Y axis DPI configuration is implemented to match the ratbag
API, however LOD is not. Whilst the internal profile structure for LOD
is handled it is not presented to any higher level interfaces. It will
however preserve the existing values configured, such that if the
Logitech software is used to configured a specific value it will be kept
when DPI values are modified by ratbag.

Button mapping and remapping is implemented based on the existing
functionality already present in the hidpp20 driver.

Report rate configuration is present on a per profile basis, however the
profile allows for wired and wireless rates to be configured
independently. The ratbag API does not have any mechanism to handle
this, so this change allows for modifying the profile report rate based
on the maximum capabilities of the device. Such that when setting the
report rate it will set both wired and wireless to the desired value
however if that value is above capability of that connection mode it
will be capped. This is useful to be able to configure the wireless
report rate (e.g. 8000Hz) of the device which can be higher than the
wired rate (e.g. 1000Hz)

In order to handle this new profile, the format_id of the profile is
persisted to the hidpp20_profile structure. This allows for the
conditional handling of the parse/write behaviour due to the significant
differences in profile binary structure. Other values are also added to
the hidpp20_profile structure to handle the additional configuration
available in this profile format.

These changes were validated with the G Pro X Wireless Superlight 2,
G203 and G900 due to changes to the common internal profile structure.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>