driver-hidpp20: handle decoding/encoding onboard profile name
Devices that support onboard profiles also allow for the name of the
profile to be configured. With recent devices and with the G-Hub
software these devices have the profile name encoded in UTF16-LE. This
change parses the raw data contained in the internal profile structure
from UTF16-LE in order to provide the profile name in UTF8. The reverse
is also handled to encode the UTF8 from the profile name into the
internal profile. This allows for displaying and updating the profile
name for onboard profiles.

Depending on the device, default (ROM) profiles may have no name stored
in the profile field. The G-Hub software will treat this as a default
name (e.g. "Profile 1"). However it is not clear if all profile formats
for all devices support the name field and if under all cases this is
encoded in UTF16-LE. As such if no name is encoded in the profile it is
treated as not having profile name support and will not allow it to be
updated when committing the onboard profile. The profile name can
initially be set in G-Hub, after which they can be displayed and or
modified within ratbag.

This has been tested with the following devices:

  - Logitech G203 - profile format id = 0x4
  - Logitech G900 - profile format id = 0x3
  - Logitech G-Pro X SL2 - profile format id = 0x6

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>