hidpp20: Add support for Extended Adjustable DPI (0x2202)
Newer Logitech devices such as the G Pro X Wireless Superlight 2 have
extended DPI adjustable support to provide access to per axis (X/Y) and
lift off distance configuration. This support is enabled via the
Extended Adjustable DPI feature (0x2202).

This change implements the necessary support to read the capabilities of
the sensor (dual axis/lift off distance) as well as the
functions/commands associated with getting the current DPI and setting
the current DPI.

The ratbag API does not currently provide a mechanism for lift off
distance configuration, because of this it defaults to configuring the
value as 2 (Medium) for all cases.

These changes were validated with the G Pro X Wireless Superlight 2.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>