hidpp20: Add support for Extended Adjustable Report Rates (0x8601)
Newer Logitech devices such as the G Pro X Wireless Superlight 2 have
extended report rate capabilities to provide access support for rates
exceeding 1000Hz. This support is enabled via the Extended Adjustable
Report Rate feature (0x8601).

This change implements the necessary support to read the capabilities of
the device as well as the functions/commands associated with getting the
current rate and setting the current rate.

The new feature provides per-connection rate capabilities and
configuration, this is currently limited to wired and wireless
connection modes. These changes capture the capabilities for both
separately. The hidpp20 driver does not currently know whether it is
being accessed via a specific connection type, due to this only the
wired case of get/set is handled. This feature is not used directly when
onboard profiles are supported.

These changes were validated with the G Pro X Wireless Superlight 2.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>