hidpp20: Clear modifiers when setting keyboard remapped button
The RATBAG_BUTTON_ACTION_TYPE_KEY type remappings do not have any key
modifiers (Ctrl/Alt/Mod/etc.) provided or remapped, as that is handled
by the macro action type (RATBAG_BUTTON_ACTION_TYPE_MACRO). However the
hidpp20 driver handles the button remapping support by using the same
mechanism as the macro implementation, which does provide support for
modifiers. However this code path does not set or clear the modifier
flags, resulting in the mapping retaining the value from the existing
configuration in the profile. This is typically the mouse button bitmask
(lower byte of the buttons field), which results in undesired behaviour.

This change clears the modifier_flags field when setting the value for
the key code.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>