hidpp20: Add support for the 0x07 profile type in G Pro X/2 devices
The G Pro X Superlight 2 and G Pro 2 Lightspeed devices received a
firmware update on 2025-04-01 to add support for a BHOP Mode. This mode
requires storing a timeout value in the profile data and as such
resulted in a different profile type (0x07). This profile is identical
to the existing 0x06 profile with a reserved field being re-purposed for
the BHOP mode timeout value.

This change implements support for this new profile type (0x07).
Additionally the existing profile type is renamed to align better with
the various devices that implement the existing 0x06 and newer 0x07
profile types. The naming of "GPX2" indicates "G Pro X or 2".

The BHOP mode has no equivalent functionality within ratbag to configure
this, so the driver only preserves the value between parse/write
operations. In the case of the existing 0x06 profile type, the default
value of 0xff is preserved.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>