data: Add device data for Logitech G Pro 2 Lightspeed
The Logitech G Pro 2 Lightspeed has a similar device data configuration
to the previous generation devices and devices of the same generation
(e.g. G Pro X Superlight 2), using the hidpp20 driver.

Support for the wireless receiver relies on the kernel supporting the
USB device within the hid-logitech-dj driver to expose the hidraw device
specific to the mouse.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>